(Refactor) Finish moving all constants to central constants module
