+++ trading/algorithms/jenetic_segmentation_oscillatory_heuristics.py
@@ -4,9 +4,10 @@ from decimal import Decimal
 from bson import ObjectId
 
 from trading.algorithms.base import Strategy
-from trading.broker import SIDE_BUY, SIDE_SELL, SIDE_STAY, PRICE_ASK_CLOSE, PRICE_ASK_HIGH, PRICE_ASK_LOW, PRICE_ASK
-from trading.broker.constants import GRANULARITY_TEN_MINUTE
-from trading.constants.constants.interval import INTERVAL_FORTY_CANDLES
+from trading.constants.price_data import PRICE_ASK, PRICE_ASK_CLOSE, PRICE_ASK_HIGH, PRICE_ASK_LOW
+from trading.constants.order import SIDE_BUY, SIDE_SELL, SIDE_STAY
+from trading.constants.granularity import GRANULARITY_TEN_MINUTE
+from trading.constants.interval import INTERVAL_FORTY_CANDLES
 from trading.indicators.misc import calc_chandalier_exits
 from trading.indicators.overlap_studies import calc_moving_average
 from trading.indicators.price_transformation import calc_standard_deviation
