+++ trading/broker/__init__.py
@@ -4,24 +4,6 @@ from trading.broker.oanda import OandaBroker
 
 MarketOrder = namedtuple('MarketOrder', ('instrument', 'units', 'side', 'type', 'price', 'expiry'))
 
-SIDE_BUY = 'buy'
-SIDE_SELL = 'sell'
-SIDE_STAY = 'stay'
-
-ORDER_LIMIT = 'limit'
-ORDER_STOP = 'stop'
-ORDER_MARKET = 'market'
-
-
-#--------------------------------#
-PRICE_ASK = 'ask'
-PRICE_ASK_CLOSE = 'closeAsk'
-PRICE_ASK_OPEN = 'openAsk'
-PRICE_ASK_HIGH = 'highAsk'
-PRICE_ASK_LOW = 'lowAsk'
-VOLUME = 'volume'
-#--------------------------------#
-
 
 BROKERS = {
     OandaBroker.name: OandaBroker
