+++ trading/constants/classifier.py
@@ -1,0 +1,1 @@
+STRATEGY_DECISION = 'decision'
