+++ trading/constants/granularity.py
@@ -1,5 +1,3 @@
 GRANULARITY_DAY = 'D'
 GRANULARITY_TEN_MINUTE = 'M10'
 GRANULARITY_HOUR = 'H1'
-
-COUNT_FORTY = 40
