+++ trading/constants/order.py
@@ -1,3 +1,6 @@
-SIDE_SELL = 'sell'
 SIDE_BUY = 'buy'
+SIDE_SELL = 'sell'
 SIDE_STAY = 'stay'
+ORDER_LIMIT = 'limit'
+ORDER_STOP = 'stop'
+ORDER_MARKET = 'market'
