+++ trading/constants/price_data.py
@@ -1,0 +1,6 @@
+PRICE_ASK = 'ask'
+PRICE_ASK_CLOSE = 'closeAsk'
+PRICE_ASK_OPEN = 'openAsk'
+PRICE_ASK_HIGH = 'highAsk'
+PRICE_ASK_LOW = 'lowAsk'
+VOLUME = 'volume'
